una locura

+++ Gravitar-clone/caracteres.c
@@ -5,7 +5,7 @@ void texto_a_caracteres(char *texto, double posx, double posy, double ang, SDL_R
 
 }
 
-const int caracter_a[7][2] = {
+const float caracter_a[7][2] = {
 	{0, 0},
 	{0, 4},
 	{2, 6},
@@ -15,7 +15,7 @@ const int caracter_a[7][2] = {
 	{0, 2}
 };
 
-const int caracter_b[12][2] = {
+const float caracter_b[12][2] = {
 	{0, 3},
 	{0, 6},
 	{3, 6},
@@ -30,14 +30,14 @@ const int caracter_b[12][2] = {
 	{3, 3},
 };
 
-const int caracter_c[4][2] = {
+const float caracter_c[4][2] = {
 	{4, 0},
 	{0, 0},
 	{0, 6},
 	{4, 6},
 };
 
-const int caracter_d[7][2] = {
+const float caracter_d[7][2] = {
 	{0, 0},
 	{0, 6},
 	{2, 6},
@@ -47,7 +47,7 @@ const int caracter_d[7][2] = {
 	{0, 0}
 };
 
-const int caracter_e[7][2] = {
+const float caracter_e[7][2] = {
 	{4, 0},
 	{0, 0},
 	{0, 3},
@@ -57,7 +57,7 @@ const int caracter_e[7][2] = {
 	{4, 6},
 };
 
-const int caracter_f[6][2] = {
+const float caracter_f[6][2] = {
 	{0, 0},
 	{0, 3},
 	{3, 3},
@@ -66,7 +66,7 @@ const int caracter_f[6][2] = {
 	{4, 6},
 };
 
-const int caracter_g[7][2] = {
+const float caracter_g[7][2] = {
 	{2, 2},
 	{4, 2},
 	{4, 0},
@@ -76,7 +76,7 @@ const int caracter_g[7][2] = {
 	{4, 4},
 };
 
-const int caracter_h[6][2] = {
+const float caracter_h[6][2] = {
 	{0, 0},
 	{0, 6},
 	{0, 3},
@@ -85,7 +85,7 @@ const int caracter_h[6][2] = {
 	{4, 0},
 };
 
-const int caracter_i[6][2] = {
+const float caracter_i[6][2] = {
 	{0, 0},
 	{4, 0},
 	{2, 0},
@@ -94,14 +94,14 @@ const int caracter_i[6][2] = {
 	{0, 6},
 };
 
-const int caracter_j[4][2] = {
+const float caracter_j[4][2] = {
 	{0, 2},
 	{2, 0},
 	{4, 0},
 	{4, 6},
 };
 
-const int caracter_k[6][2] = {
+const float caracter_k[6][2] = {
 	{0, 0},
 	{0, 6},
 	{0, 3},
@@ -110,13 +110,13 @@ const int caracter_k[6][2] = {
 	{3, 0},
 };
 
-const int caracter_l[3][2] = {
+const float caracter_l[3][2] = {
 	{4, 0},
 	{0, 0},
 	{0, 6},
 };
 
-const int caracter_m[5][2] = {
+const float caracter_m[5][2] = {
 	{0, 0},
 	{0, 6},
 	{2, 4},
@@ -124,14 +124,14 @@ const int caracter_m[5][2] = {
 	{4, 0},
 };
 
-const int caracter_n[4][2] = {
+const float caracter_n[4][2] = {
 	{0, 0},
 	{0, 6},
 	{4, 0},
 	{4, 6},
 };
 
-const int caracter_o[5][2] = {
+const float caracter_o[5][2] = {
 	{4, 0},
 	{0, 0},
 	{0, 6},
@@ -139,7 +139,7 @@ const int caracter_o[5][2] = {
 	{4, 0},
 };
 
-const int caracter_p[5][2] = {
+const float caracter_p[5][2] = {
 	{0, 0},
 	{0, 6},
 	{4, 6},
@@ -147,7 +147,7 @@ const int caracter_p[5][2] = {
 	{0, 3},
 };
 
-const int caracter_q[9][2] = {
+const float caracter_q[9][2] = {
 	{2, 2},
 	{3, 1},
 	{2, 0},
@@ -159,7 +159,7 @@ const int caracter_q[9][2] = {
 	{4, 0},
 };
 
-const int caracter_r[7][2] = {
+const float caracter_r[7][2] = {
 	{0, 0},
 	{0, 6},
 	{4, 6},
@@ -169,7 +169,7 @@ const int caracter_r[7][2] = {
 	{4, 0},
 };
 
-const int caracter_s[6][2] = {
+const float caracter_s[6][2] = {
 	{0, 0},
 	{4, 0},
 	{4, 3},
@@ -178,27 +178,27 @@ const int caracter_s[6][2] = {
 	{4, 6},
 };
 
-const int caracter_t[4][2] = {
+const float caracter_t[4][2] = {
 	{2, 0},
 	{2, 6},
 	{4, 6},
 	{0, 6},
 };
 
-const int caracter_u[4][2] = {
+const float caracter_u[4][2] = {
 	{0, 6},
 	{0, 0},
 	{4, 0},
 	{4, 6},
 };
 
-const int caracter_v[3][2] = {
+const float caracter_v[3][2] = {
 	{0, 6},
 	{2, 0},
 	{4, 6},
 };
 
-const int caracter_w[5][2] = {
+const float caracter_w[5][2] = {
 	{0, 6},
 	{0, 0},
 	{2, 2},
@@ -206,7 +206,7 @@ const int caracter_w[5][2] = {
 	{4, 6},
 };
 
-const int caracter_x[5][2] = {
+const float caracter_x[5][2] = {
 	{0, 0},
 	{4, 6},
 	{2, 3},
@@ -214,7 +214,7 @@ const int caracter_x[5][2] = {
 	{4, 0},
 };
 
-const int caracter_y[5][2] = {
+const float caracter_y[5][2] = {
 	{0, 6},
 	{2, 4},
 	{4, 6},
@@ -222,14 +222,14 @@ const int caracter_y[5][2] = {
 	{2, 0},
 };
 
-const int caracter_z[4][2] = {
+const float caracter_z[4][2] = {
 	{0, 6},
 	{4, 6},
 	{0, 0},
 	{4, 0},
 };
 
-const int caracter_0[5][2] = {
+const float caracter_0[5][2] = {
 	{4, 0},
 	{0, 0},
 	{0, 6},
@@ -237,12 +237,12 @@ const int caracter_0[5][2] = {
 	{4, 0},
 };
 
-const int caracter_1[2][2] = {
+const float caracter_1[2][2] = {
 	{2, 0},
 	{2, 6},
 };
 
-const int caracter_2[6][2] = {
+const float caracter_2[6][2] = {
 	{0, 6},
 	{4, 6},
 	{4, 3},
@@ -251,7 +251,7 @@ const int caracter_2[6][2] = {
 	{4, 0},
 };
 
-const int caracter_3[7][2] = {
+const float caracter_3[7][2] = {
 	{0, 0},
 	{4, 0},
 	{4, 3},
@@ -261,7 +261,7 @@ const int caracter_3[7][2] = {
 	{0, 6},
 };
 
-const int caracter_4[5][2] = {
+const float caracter_4[5][2] = {
 	{0, 6},
 	{0, 3},
 	{4, 3},
@@ -269,7 +269,7 @@ const int caracter_4[5][2] = {
 	{4, 0},
 };
 
-const int caracter_5[6][2] = {
+const float caracter_5[6][2] = {
 	{0, 0},
 	{4, 0},
 	{4, 3},
@@ -278,7 +278,7 @@ const int caracter_5[6][2] = {
 	{4, 6},
 };
 
-const int caracter_6[5][2] = {
+const float caracter_6[5][2] = {
 	{0, 6},
 	{0, 0},
 	{4, 0},
@@ -286,13 +286,13 @@ const int caracter_6[5][2] = {
 	{0, 3},
 };
 
-const int caracter_7[3][2] = {
+const float caracter_7[3][2] = {
 	{0, 6},
 	{4, 6},
 	{4, 0},
 };
 
-const int caracter_8[7][2] = {
+const float caracter_8[7][2] = {
 	{0, 3},
 	{0, 6},
 	{4, 6},
@@ -302,7 +302,7 @@ const int caracter_8[7][2] = {
 	{4, 3},
 };
 
-const int caracter_9[5][2] = {
+const float caracter_9[5][2] = {
 	{4, 0},
 	{4, 6},
 	{0, 6},
@@ -310,7 +310,7 @@ const int caracter_9[5][2] = {
 	{4, 3},
 };
 
-const int caracter_derecha[5][2] = {
+const float caracter_derecha[5][2] = {
 	{0, 3},
 	{12, 3},
 	{9, 5},
@@ -318,7 +318,7 @@ const int caracter_derecha[5][2] = {
 	{12, 3},
 };
 
-const int caracter_izquierda[5][2] = {
+const float caracter_izquierda[5][2] = {
 	{12, 3},
 	{0, 3},
 	{3, 5},
@@ -326,7 +326,7 @@ const int caracter_izquierda[5][2] = {
 	{0, 3},
 };
 
-const int caracter_arriba[5][2] = {
+const float caracter_arriba[5][2] = {
 	{5, -2},
 	{5, 10},
 	{3, 7},
@@ -334,7 +334,7 @@ const int caracter_arriba[5][2] = {
 	{5, 10},
 };
 
-const int caracter_abajo[5][2] = {
+const float caracter_abajo[5][2] = {
 	{5, 10},
 	{5, -2},
 	{3, 1},
@@ -342,7 +342,7 @@ const int caracter_abajo[5][2] = {
 	{5, -2},
 };
 
-const int caracter_espacio[1][2] = {
+const float caracter_espacio[1][2] = {
 	{0, 0},
 };
